#40 - adicionada verificacao de login
