+++ backend/api/routers/auth_routes.py
@@ -1,9 +1,9 @@
 from api.database.config import Session, engine
-from api.schemas.users import Token
+from api.schemas.auth import Token, LoginInput
 from api.models.users import User
 from fastapi import APIRouter, HTTPException
 from datetime import datetime, timedelta
-from api.utils.auth_services import create_access_token
+from api.utils.auth_services import create_access_token, verify_password
 
 ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
@@ -19,15 +19,24 @@ router = APIRouter(
 session = Session(bind=engine)
 
 
-@router.post("/token/{id}", response_model=Token)
-async def getUserAccessToken(id: int):
-    user = session.query(User).filter(User.id == id).first()
+@router.post("/token", response_model=Token)
+async def getUserAccessToken(data: LoginInput):
+    user = session.query(User).filter(User.username == data.username).first()
     if user == None:
         raise HTTPException(
             status_code=401,
             detail="Incorrect username or password",
             headers={"WWW-Authenticate": "Bearer"},
         )
+
+    password_match = verify_password(data.password, user.password)
+    if not password_match:
+        raise HTTPException(
+            status_code=401,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+
     access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
     access_token = create_access_token(data={"sub": user.username}, expires_delta=access_token_expires)
     return {"access_token": access_token, "token_type": "bearer"}
