+++ backend/api/schemas/auth.py
@@ -1,0 +1,15 @@
+from typing import List, Optional, Generic, TypeVar
+from pydantic import BaseModel, Field
+from pydantic.generics import GenericModel
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    username: str | None = None
+
+class LoginInput(BaseModel):
+    username: str
+    password: str
